HUCO-166 | [feat] make game setting manager and move duplicated code to prefab
