+++ HandsUp/Assets/OpeningScene/Scripts/GameSettingManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b73a1c44259274c4b900900c816171a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
